Adding Rating/Review for Product Backend

+++ backend/src/main/java/org/example/backend/model/Review.java
@@ -1,7 +1,7 @@
 package org.example.backend.model;
 
 public record Review(
-        int ratingCount,
+        float ratingCount,
         String commentary
 ) {
 }
